+++ src/models/conversation.ts
@@ -6,7 +6,7 @@ import { QueueItem } from './queue-item';
 import { EmergencyMessageType } from '../constants/message-types/emergency';
 import { NamespacedConversation } from '../namespaces/interfaces';
 import { Store } from './store';
-import { Message, Origin, MessageState } from './message';
+import { Message, Origin, MessageState, nextUniqueId } from './message';
 import { CALL_INFO, CONTENT_TYPE, REPLY_TO } from '../constants/headers';
 import { CALL_SUB, MULTIPART_MIXED, PIDF_LO, TEXT_PLAIN, TEXT_URI_LIST } from '../constants/content-types';
 import { Multipart, MultipartPart, CRLF } from './multipart';
@@ -397,6 +397,7 @@ export class Conversation {
     const message: Message = {
       // if no message id is specified, use the internal sequence
       id: messageId ?? this._messageId++,
+      uniqueId: nextUniqueId(),
       origin: Origin.LOCAL,
       conversation: this,
       dateTime: new Date(),
@@ -562,6 +563,7 @@ export class Conversation {
 
       this._addNewMessage({
         id: this.mapper.getMessageIdFromHeaders(callInfoHeaders) as string,
+        uniqueId: nextUniqueId(),
         origin,
         conversation: this,
         dateTime: now,