+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This should limit memory consumption if ng112-js is used in long running applica
 - Property `uniqueId` in messages that is unique across all messages and conversations within ng112-js
 - Auto-close of open calls on agent disposal
 - Configurable grace period for agent disposal
+- More detailled logging on message errors.
 
 ---