+++ CHANGELOG.md
@@ -22,6 +22,10 @@ This should limit memory consumption if ng112-js is used in long running applica
 
 ---
 
+## [0.11.21] - 2021-04-24
+### Added
+- Missing export for `AgentState`
+
 ## [0.11.19] - 2021-04-19
 ### Added
 - JsSIP version in UserAgent string